Updated openstack/openstack Project: openstack/swift  c0640f87107d84d262c20bdc1250b805ae8f9482
SLO: Make etag and size_bytes fully optional

Previously, we still required that clients send "etag" and "size_bytes"
keys in their segment definitions. This was done as a way to guard
against typos leading to an accidental lack of verification.

However, typos should already be caught when we check for extra keys. As
a result, the only truly required key is "path".

Change-Id: Ie1d8691115f8c68b5a3f3b59317cdab59f9a3fca