Updated openstack/openstack Project: openstack/nova  146b27f22327f8d60c1017c22ccf18d0e16f1eb7
Make test_create_delete_server_with_instance_update deterministic

We're seeing race failures in this test because the ip_addresses
can show up in an unexpected notification depending on when
allocate_for_instance in the neutron network API completes and
the instance.info_cache.network_info data gets stored in the
database.

We can resolve the race by using SpawnIsSynchronousFixture which
makes the allocate_for_instance network API a blocking call
until the network_info is returned, and by that time it's stored
in the instance_info_cache in the database which is where the
versioned notification pulls it from in _send_versioned_instance_update.

Change-Id: Id482220b8332549a07efb4f82212d74e6e7b9d6c
Closes-Bug: #1649403