Updated openstack/openstack Project: openstack/keystone  55755e3bf7c461ceee2bc033657eefb4060453f2
Add id to conflict error if caused by duplicate id

This patch adds test coverage for duplicate erros on objects that
would raise an exception conflict by duplicate ids. It also adds
the id of the object in place of the name in the duplicate error
message if a name field for the object cannot be found.

Closes-Bug: #1599296

Change-Id: Ief9ef3d29ee9ac2da1c205247601399fb6f79d7b