Updated openstack/openstack Project: openstack/diskimage-builder  339ecee2b20e88f348b8e9fffc26cadc204f82a5
elements: dib-python: Add python2 as the default version for openSUSE

Commit 6278371eaa13("Make dib-python use the default python for distro")
added default python version for various distros but it missed openSUSE
which leads to build failures since the openSUSE elements are pulling
python2 packages. Add openSUSE to the list of python2 distributions
until python3 support for the openSUSE elements is in place.

Change-Id: I95f1fa849a22607c430387a2a915f9d19c9c209f