Updated openstack/openstack Project: openstack-infra/release-tools  8b754f82be07abecc3047ec5211ec13fcd4b154a
Handle non-python libraries in announce.sh

Some libraries are *NOT* python projects and as such do not have a
setup.py file.  This means that we don't generate announcement emails
for these projects.  A recent example is available at [1].

This change detects that situation and works with the information
available to ensure the release is announced.

[1] http://logs.openstack.org/bf/bf85b2cf8bfc4c2456f9666d1ceaed3cd0d34e32/release/puppet-tripleo-announce-release/a90a40c/console.html#_2016-11-15_04_33_50_485019

Change-Id: I5b475ea001e9a2f13fb413007f96bfd39c8ab439