Updated openstack/openstack Project: openstack/glance_store  32bdeddfb101c33df1058e4a1d51799953059ede
Use storage_url in DB for multi-tenant swift store

When using swift multi-tenant store, using a user's service
catalog to obtain the storage url for an object does not work
if the image was not created by the same user. This is because
the user's context, if used to establish a connection, would only
be looking at a location url formed with the wrong owner information
resulting in an object GET failure (404).

This patch solves this issue by a 'revert in principle' to the old code
logic where the storage_url for an image when using multi-tenant store
was fetched from the database via 'location.swift_url'. The change
that had introduced the fetch of storage_url from user's context is [1].

[1] I7bc23dfc11900b55f45fe98144d14f883c381c9f

Co-Authored-By: Hemanth Makkapati <hemanth.makkapati@rackspace.com>
Co-Authored-By: Dharini Chandrasekar <dharini.chandrasekar@intel.com>

Partial-Bug: #1625075

Change-Id: I8c22a0ab566cf1ec99cbee78ea43ef28abcd8dc0