Updated openstack/openstack Project: openstack/python-glanceclient  8b96bbc34cc0a06f5eb540d42dcfb73229329011
Revert "Add last_request_id member to HTTPClient and SessionClient"

This reverts commit 9e532db8b0f0ba537edef143a6f5380a2aaa1e4b.

If glanceclient is used in multi-threaded environment, then there is a
possibility of getting invalid/wrong last request-id. To avoid this,
need to use thread local storage to store last-request-id and add
public method to return this request-id to caller.

http://specs.openstack.org/openstack/openstack-specs/specs/return-request-id.html#alternatives

Change-Id: I08d8d87fc0cc291f1b930b2c0cfc110ec8394131