Updated openstack/openstack Project: openstack/ironic  5b16eb99f22ec556559a5318c4b4f043d5b74b15
Reusing oneview_client when possible

A simple experiment showed that all calls to OneView were calling for
authentication. Reviewing the code, I perceived that every single call to
python-oneviewclient were creating a new instance of the client, which
authenticated again and generated a new session token on OneView.

This patch aims to reuse oneview_client instances when possible, avoiding
to create unecessary new sessions that can be reused.

Co-Authored-By: Hugo Nicodemos <nicodemos@lsd.ufcg.edu.br>
Co-Authored-By: Xavier <marcusrafael@lsd.ufcg.edu.br>

Change-Id: I0c26f0e033ca7f53f648c4ed29c6ef19f6f2e578
Closes-Bug: #1615070