Updated openstack/openstack Project: openstack/nova  0415ee80b5c8354124290ac933a34823f2567800
Merge v20 view builder for limit API

Now v2 and v2.1 code is merged and we do not need
to maintain the different set of view builder.
Previously there were different way of building the
complete response for limit API. v2 used to do with extension
and v2.1 with view builder itself.

Change-Id: I79fc803f471c9810c413d1b2983dc13763fa4633