Updated openstack/openstack Project: openstack-infra/shade  1be8d9600c145bd69a0b2ca87a0d53f68667d5e5
Plumb nat_destination through for ip_pool case

Currently, if ip_pool is provided to add_ips_to_server, nat_destination
is not passed along to the things that need it. Fix that. Also, while
in that, pass server to create_floating_ip so that the better neutron
FIP creation routine gets used.

Story: #2000820
Change-Id: I32fd485cb7e7064eaae35578a3643100ee85a096