Updated openstack/openstack Project: openstack/openstack-doc-tools  e7f3aee59363ac962bec5171d2b1790505929067
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove openstack-doc-tools.

Change-Id: I3b7f2f7a18325a9cb0e461e9bac2648c0c27bf4c