Updated openstack/openstack Project: openstack-infra/tripleo-ci  b9bde713738952bc3788518dd2ac9d73b1ada2f2
Default libvirt_cpu_mode to none when using qemu

An upcoming update to qemu-kvm (2.6.0) with CentOS/RHEL 7.3 can break
existing implementations where qemu is the configured hypervisor and
'host-model' is used.
As such, ensure libvirt_cpu_mode is set to none when qemu is in use.

Worth noting that devstack has been setting 'none' [1] as well.

[1]: https://github.com/openstack-dev/devstack/blob/1c442eebc8fe005af453bd610e750a1919a2b3ed/lib/nova_plugins/hypervisor-libvirt#L42

Change-Id: I4a7c3ced59c46fb2582122260922813b09a3d042