Updated openstack/openstack Project: openstack/heat  8fc348910857ad0385fe6c5d884bda4ead5f9591
Do not set the mime-type for userdata

We set the mime-type to  'x-shellscript' for all
single userdata when user_data_format='SOFTWARE_CONFIG',
it's incorrect, for example, if user create a server
with a cloud-config userdata, we will set the mime-type to
'x-shellscript', then cloud-init will consume the userdata
as a shellscript and fail to execute.

This change will improve this by not setting the mime-type
and letting cloud-init figure it out from the first line
in the file.

Closes-Bug: #1643739
Change-Id: If026884a95c58bb81a0f04ab415909011b1a5414