Updated openstack/openstack Project: openstack/requirements  0d40dd25d5a531c8838d5172f96e3e134a7295f2
Add trio2o to the list of projects

The requirements in Trio2o can not be updated by the bot
yet, but the "check-requirements" job has already been added
to project-config/blob/master/zuul/layout.yaml.
So add trio2o project to projects.txt to trigger the
requirements sync. bot

Change-Id: I36cecb552b1d8129241d2c25fb4ac532de2c5ab9
Signed-off-by: joehuang <joehuang@huawei.com>