Updated openstack/openstack Project: openstack/neutron  d1b31485eba0b4f236d755c6e64336b142eb8c72
QoS: update the database before notifying the backend on delete

Previously, the QoS plugin was executing the DB delete operation
after notifying the backend about the operation. This leads
to situations where the backend is notified for deletion, and
then the policy can't be deleted because it's in use.

Correct order is, DB delete first, then notify the backend.

Change-Id: I357543832b9359bf169d05d079bd153f0ee591c4
Closes-Bug: #1646370