Updated openstack/openstack Project: openstack/nova  b27d091fb0a1305e4de6ba8702186e06f10fa99c
Stop allowing tags as empty string

Nova support tagging of instance or other resource like device etc.
But those tags are allowed to be as empty string which does not make
much sense or any real use case.

Updating single tag with empty string is not possible as it will
be 404 due to url itself has single tag as id.

But updating all tags with few or all of them as empty string
does not complain. In those cases, empty tags are being accepted and
stored as same.

Main issue with those is that empty tag cannot be deleted/show as single
tag as it will again 404 with url not found. Only way to delete/show
such tag is to delete all together.

Empty tag should not be allowed at first.
This commit makes empty string request to 400 which are
currently 200. Doing for server and device tags and same can
be adopted for other resource tagging also if any in future.

Fixing this as bug fix hoping no backward incompatibility
logically. If such tag could have been deleted alone then we could have
fix this with microversion.

Change-Id: I18a81f19205b2a40ca470067a9576f2f72ff0f13
Closes-Bug: #1648663