Updated openstack/openstack Project: openstack/oslo.messaging  8d2331a3b7b65d15adc672b25a13e0f7283c7b96
tox: rename zeromq target

The current zeromq target name are too long and reach the limitation
of 127 chars of the sheband of the virtualenv binary.

A target name must not be more than 22 chars to run in gate with
the tox job template.

This renames them to a shorter one.

It also creates aliases for the legacy jobs that we will remove
when the tox jobs are ready.

Change-Id: I5cbc368e7ea252f8c0c949c8e7ee3a70a560db22