Updated openstack/openstack Project: openstack/api-wg  e01a677e96d36c7885e262a207fbb66c23e18b99
Add clarification on 400 vs. 404 guideline

There was some confusion of the appropriateness of using 400 when a
resource referenced in the body of a request did not exist. An example
of this is when creating a VM, but passing a non-existent flavor in the
request body. This patch clarifies that usage, and adds language to
emphasize that it is critical to be explicit as to exactly what is wrong
so that the issuer of the bad request knows exactly what must be done to
correct the problem.

Change-Id: Ie7b3f4cc49056380f31796aa4cebd99a574c132a