Updated openstack/openstack Project: openstack/horizon  fc7f82400e6d6c499e81a57a0762106288890111
Update for ui bootstrap changes

UI Bootstrap decided to prefix all of their services/functions, so we
need to add these prefixes for the code to continue to function.

This patch primarily addresses functionality - some styling is broken
even with this patch, but those breakages can be fixed once this
patch lands.

Change-Id: Ia6f593e27f81b2b00b6376426cc469a776c4add7
Depends-On: I7d8ebd4e9018332c1974672a0b99080c152ce6c3