Updated openstack/openstack Project: openstack-infra/shade  0784919852f2d42b626482c8d1e9ad214441b3f5
Convert glance parts of task test to requests_mock

Remove the mocking of _image_client and go straight for requests_mock.

It's worth noting doing this caught a bug in the POST to /tasks where we
were using data= instead of json=. Yay tests.

Also - we've added a failure in the task poll loop. The first GET for
the task status returns a 503 which is then appropriately skipped.

Change-Id: Ia50333d8a1c62b68e89f08aafe4b7e6caee8ef5a