Updated openstack/openstack Project: openstack-infra/shade  642554b41abd6000e2f9d3337f537908cff9674d
Pass md5 and sha256 to create_object sanely.

The logic to skip md5/sha256 calculation was flawed, as it pulled the
wrong keys from the image_kwargs. The whole idea of pulling parameters
from the image create kwargs is too clevel though. Just go ahead and
pass the values as real python parameters.

Change-Id: I516b0143d6874bd31f5050fc6bee0dc8227dfcda