Updated openstack/openstack Project: openstack/nova  8e170ec3b410b1881edaa3d8becd4292f1c03243
libvirt: Detach volumes from a domain before detaching any encryptors

I983f80822a5 introduced retry logic around the detaching of volumes from
a libvirt domain. However it also reordered the surrounding code so that
we attempt to detach any encryptors from a volume first while the volume
is still attached to the domain. This can lead to `Device or resource
busy` errors if the volume is still being used by the instance when we
attempt to detach the encryptors.

Closes-bug: #1642628
Change-Id: Ia0f8e725ec8a0fbc44bd4592b021dea978cf4e4f