Updated openstack/openstack Project: openstack/keystone  c46ff2b1abcddeb144633c841143b1d5316a1032
expose v3policy failure with is_admin_token

The cloud_admin rule is incorrectly written, and will result in
a TypeError when any action is taken that uses the cloud_admin
rule (most of our APIs). This change exposes that error by
attempting to load and enforce the policy, this is done for
both sample policies we provide.

Partial-Bug: 1547684

Change-Id: If039bfb13741ee5a4b6b9c34b45ceafcfb5b342e