Updated openstack/openstack Project: openstack/ironic  4c0653a98d52e4120da192983c787cc4c4d8798c
Make CONF.debug also reflect on IPA

Prior to this patch, setting the CONF.debug to True in the ironic.conf
doesn't enable the debug logs in IPA. The user had to also append
"ipa-debug=1" to the CONF.pxe.pxe_append_params configuration option.

This patch is changing this behavior by making the CONF.debug option
reflect on the IPA logs as well. As I understand, IPA is just an
extension of the Ironic code that runs in-band on the node allowing
Ironic to access the local hardware. This way, in order to enable debug
as a whole the user only needs to set one configuration option.

The code in this patch checks if CONF.debug is set to True and also
check if "ipa-debug" is not explicit set in CONF.pxe.pxe_append_params
configuration - to not overwrite the user configuration - and only then
will append "ipa-debug=1" to it.

The "ipa-debug=1" kernel parameter was also removed from the devstack
lib because CONF.debug is always set to True there.

Change-Id: I675fb11248d3789d45cf5905b7c15368f026b345
Closes-Bug: #1649561