Updated openstack/openstack Project: openstack/ironic  d2a02881a7532b1e34cb3be0bcaa8d82c8d5451b
DevStack: Make $IRONIC_IMAGE_NAME less dependent of the name in DevStack

This patch is making the $IRONIC_IMAGE_NAME less dependent on the full
name of the image in DevStack. We will need different images for the
UEFI work and this would break it so, let's make it a bit more smart.

Change-Id: Iac8668e9470bc151c10ea145d4394ba7ff7041d8