Updated openstack/openstack Project: openstack/oslo.policy  98bc1354e2f8a478a21c803ffff2b8a54de8c115
Remove dead code and use default value of argparse.

In the 'tool' method, the var 'apply_rule' was tested, and the
method returned if it was true. Then, it was tested again even
though it could only be 'false' (or not be None).
Also, an unused parameter of main() had been removed.

Closes-Bug #1650599
Change-Id: I4b265ef609d4a5fc8128f40359cf8d04ee8cbe28
Signed-off-by: Sami Makki <mail@samimakki.fr>