Updated openstack/openstack Project: openstack/cinder  cd06286062855039b524512d7c56d200f83705a9
Rolling upgrade procedure documentation

One of the requirements to achieve assert:supports-rolling-upgrade TC
tag is to provide a "plan that allows operators to roll out new code to
subsets of services, eliminating the need to restart all services on new
code simultaneously".

This commit documents such rolling upgrade procedure in the devref,
which is last outstanding requirement for Cinder.

Change-Id: I562deafcd4bb9f5666cc35ae77a6bf20ab889612