Updated openstack/openstack Project: openstack/glance  38d42ad3c61b8948dfe8ef239db78b363d92b8ef
Remove v3 stub controller

This patch removes the stub controller for v3 that was introduced
in Mitaka when Glare became a standalone service.  As the v3 API
was always EXPERIMENTAL, this removal does not require a deprecation
period.  (Also, anyone interested in artifacts is using Glare, not
the v3 API.)

Change-Id: Ie16ba13c5970596eebf7b757c0712358c80f376e
Closes-bug: #1650108