Updated openstack/openstack Project: openstack/nova  f5ef0f3e0f06bb1a03b5bbaea81f31c3e3b1a843
Make nova-manage emit a traceback when things blow up

If we hit an unhandled error in nova-manage we don't
get a traceback currently which makes debugging the actual
failure hard. This change emits a traceback in main() when
we get an unhandled exception up from one of the subcommands.

Change-Id: Id8b8ad3c97c6834abc4290319cd6f2cd5aafa479
Related-Bug: #1650694