Updated openstack/openstack Project: openstack/requirements  eea4e52ead9f370f83f3bd8fcc23a9bdea94aaf4
bump minimum keystonemiddleware to 4.12.0

This is part of a larger effort to solve the issue of long running
operations timing out when a token expires. The change to allow
an expired token to be validated was introduced in 4.12.0. The
server side code has already been implemented in ocata-2.

see bp allow-expires

Change-Id: I2e23c6bb7d51ee1486693bb40d3c6a3a8b2cce4f