Updated openstack/openstack Project: openstack/trove  d9580fccbea62a9c0c443d1638ee4742bcd3683b
Avoid double-check when waiting for test states

Sometimes tests fail becaused there's a double check for
the instance state.  If the state changes in between the
two checks, the test will fail to acknowledge the first
'correct' response and timeout while waiting for the state
to re-occur (which doesn't happen).

Sample output would look like:
[REPORT] Polling instance '<id>' for state 'REBOOT', was 'REBOOT'.
[REPORT] Polling instance '<id>' for state 'REBOOT', was 'ACTIVE'.
...
[REPORT] Polling instance '<id>' for state 'REBOOT', was 'ACTIVE'.
...
[REPORT] Status of instance '<id>' did not change to 'REBOOT' after 1201.3s.

Also broke down the instance-create tests into instance-create and
instance-init-create so that the initialzed instance won't
need to be created for manual runs or multi runs in the gate.

Reorganized the backup create so that on fast nodes there's
more chance that the backup is still around while the
negative tests run.

Change-Id: I9731bf7a93e7b8b3398cf73d1bc15d70f5bfa96a