Updated openstack/openstack Project: openstack/cinder  323c8acd91eaa58aef99ea3300e49589ed08e76c
Add get_all capability to volume_attachments

One of the useful things that was missing from the
volume_attachments code was get_all methods.

This patch adds a get_all and a get_all_by_project, it
also goes ahead and adds some filtering capability to
the existing get_by_xxxx calls since we added the framework
for it in the get_all additions.

I also looked at refactoring our db methods for attach to just:
  * attach_create
  * attach_update
  * attach_destroy
  * attach_get
  * attach_get_all

This would probably be good as an independent effort to
clean things up and bring these calls more in line with
others, but there's a lot of work to update the objects
and existing code, might be better to wait until after
implementing the new attach API.

Co-Authored-By: Michał Dulko <michal.dulko@intel.com>

Change-Id: I40614fe702f726c74ff05f93faaf6ee79253447f