Updated openstack/openstack Project: openstack/nova  ba3c17c3ee0adf6e8de9f025f5c9dd0a1bc65d04
nova-manage: squash oslo_policy debug logging

I don't know when it started but you can see from a modern
usage of nova-manage that oslo_policy debug logging is being
emitted. We don't need that debug information in the console
for nova-manage, so this patch hides it just like we do for
oslo_db.

Change-Id: If1a8e8d473846f2626f5e01210f00ad06916a383
Related-Bug: #1650694