Updated openstack/openstack Project: openstack/tempest  4561b60254c9002294bce29895e1bcf0afd4518e
Improve help message for scheduler_available_filters

Config option scheduler_available_filters default value
'all' is not much clear as it doesn't mean that all the
existing filters are enabled. It is just the indication
of the case when all of the available filters
(from DEFAULT.scheduler_available_filters in nova.conf)
are enabled.

Also add clear doc string for is_scheduler_filter_enabled
function which is used to evaluate this option value.

Change-Id: If9a52003309686cc257e62737e52c2cc52ddc659
Closes-Bug: #1628443