Updated openstack/openstack Project: openstack-infra/project-config  3c4f0ffae6cdf0d71adb8b29d3cc642b7721883d
Fix python-heatclient docs publishing

heatclient was only publishing docs in release pipeline but used the
wrong macro - use the proper template for it so that publishing works.

If publishing happens only in release pipepline, the tags-only jobs
(used by client publish template) needs to be used to have content at
top-level as well.

Change-Id: Ia30ed578efdd3ea9d8b9f2ef37afc5d07f88dcb5