Updated openstack/openstack Project: openstack-infra/project-config  f65ff79672fe39d4fa3374209cb2d96afc9c5794
add update_constraints.sh for use by new release job

This script is based on the existing steps in release.sh for updating
the constraints list after releasing one of our Python projects. It's
placed in its own commit so it is available on the release nodes when
the new job in the following commit is merged.

Change-Id: Ie3e6d0e3501fa0cba251e000430bd0740a2bd331
Signed-off-by: Doug Hellmann <doug@doughellmann.com>